refactor: rename assets to follow convention
