+++ frontend/index.html
@@ -2,7 +2,7 @@
 <html lang="en">
   <head>
     <meta charset="UTF-8" />
-    <link rel="icon" type="image/svg+xml" href="/src/assets/wieFavicon.svg" />
+    <link rel="icon" type="image/svg+xml" href="/src/assets/favicon-wie.svg" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <title>Women in Engineering</title>
   </head>
